scripts/requirements.txt
# Inference Scripts Requirements

# Core dependencies
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.0.0

# Deep Learning
onnxruntime-gpu>=1.12.0  # Use onnxruntime for CPU-only
torch>=1.10.0
torchvision>=0.11.0

# UI
PyQt6>=6.0.0

# Configuration
PyYAML>=5.4.0

# Tracker dependencies (already in main requirements)
scipy>=1.7.0
